style: apply pint psr-12 preset across the project
Mostly mechanical: blank line after <?php, space in fn () arrow functions,
brace on its own line for class definitions, multi-line empty constructor
bodies. No behaviour change. 55 tests still green, phpstan still clean,
rector dry-run also clean.

Two follow-up rector findings applied in the same pass:

- src/Catalog.php: drop redundant (int) cast on $item->fileSize() ?? 0
  (the ?? short-circuit already produces an int).
- src/Controller.php: switch the rebuild-summary assertion from
  !== null to instanceof PackagesJsonResult for stronger type narrowing.

+++ src/Catalog.php
@@ -1,4 +1,5 @@
 <?php
+
 declare(strict_types=1);
 
 namespace Composerd;
@@ -27,14 +28,14 @@ final class Catalog
                 package: $m[2],
                 version: $m[3],
                 path: $path,
-                size: (int) ($item->fileSize() ?? 0),
-                lastModified: (int) ($item->lastModified() ?? 0),
+                size: $item->fileSize() ?? 0,
+                lastModified: $item->lastModified() ?? 0,
             );
         }
 
         usort(
             $entries,
-            fn(CatalogEntry $a, CatalogEntry $b): int => strcmp($a->path, $b->path)
+            fn (CatalogEntry $a, CatalogEntry $b): int => strcmp($a->path, $b->path)
         );
 
         $hashInput = '';
